find_work/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
Django==5.2.9
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
